refactor: move wrapper to moonshot-cache.ts to reduce extra-params.ts invasion
Reduces extra-params.ts changes from ~135 lines to ~20 lines.
All Moonshot-specific logic is now isolated in moonshot-cache.ts.
